Switch to the Inter UI font. (#125)
[Inter UI] is a common font that's used in many OSS projects like
Blender. This commit switches the Bevy editor to use it.

Even if we don't ultimately adopt this font, this PR adds the
infrastructure for custom fonts to the `Theme` structure, so it's
useful. The default Bevy font, Fira Sans Mono, isn't a good choice for
readable UI as it's monospaced.

[Inter UI]: https://rsms.me/inter/