Scroll box widget (#96)
* Implement very basic scroll_box widget

* Update spawn_scroll_box to allow populating list contents

The challenge is that spawn_scroll_box can only return the
scroll_box root node.
But we might want to add more UI inside the scroll box.
To allow this I've added a closure that will be called after the
ScrollBox (content node) is spawned.

* Rework to a less temporary scroll box implementation

Renamed components and prepare for X and Y scrolling
(currently only Y scrolling is implemented)

* Scroll box X/Y/Both axis scrolling

* Fix clippy/doc warnings/errors

* Hide scroll bar when no scrolling is needed