Launcher and Projects (#135)
* transform bevy_editor into a plugin

* fix small issues

* Allow project game to be run from the editor

There is a new callback where you can setup your game code.
This callback is called when running the game without the editor.
A new button as appear to run the game from the editor.
This button just call `cargo run -- -game` in the project directory.

* Add project management system

This module will allow users to handle project
- create new ones from templates
- keep track of the projects on the machine
- run projects

* Rename alien_cake_addict to getting_started

It's purpose is to be a simple project that guide can based themselves
on as a starting point
It will help on the development of per-project features (gameplay, ui, etc)
And would allow new user to have a simple project to start with and
experiment

* Fix oupsys

forgot some cleaning and dependencies

* Remove run button

This button was mainly for the show, and not really working
I'll bring it back someday it's a very easy feature to implement

* Add a very ugly but working launcher

>> Now it's the default workspace project
I wanted to make this prettier but was facing issues and it will be
refactored anyway so I'm just going to leave it like this for now.

* Replace EditorPlugin by an new App

I don't wanted to call this an Editor Something because people would get
confused on why they will use the EditorPlugin even in runtime mode.
This App will be configurable in code or through configuration files
(once this is implemented).

* Fix errors/warning to pass CI

* Fix clippy error on linux

* Update project info when the editor is opened

This will allow any project to be saved in cache even though it wasn't
run by the launcher.

* Remove project from project list, if invalid

When running a project from the launcher but this one doesn't exist or
isn't valid, it will be removed from the project list and on the UI

* Add small doc to the `bevy_editor_launcher` crate

* Implement running project in other platforms than windows

Hopefully that works, it should but couldn't test it

* Remove name from project info

Use the last segment of the path as the project name

* Use resource to handle the project list

Before I was loading and saving every time I wanted to change something

* Fix typos

* Fix getting_started bevy_editor dependency