Handle context menu open event base of owner z-index
If you had multiple UI element overlapping with each their own context
menu, it would spawn all of them and only the last one would be visible.
I've changed it, so only 1 can be spawn at a time, and it used the
owner/initiator z-index widget to determine which one should be spawned.